+++ Multifacile/DataFileMgr.h
@@ -30,11 +30,11 @@ public:
     DataFileMgr(const QString &fileName);
     ~DataFileMgr() { xmlFile.close(); }
 
-    bool setValue(const QString &group, const unsigned int &time, const bool &noError = false, const unsigned int &table = 0);
-    const QString value(const QString &group, const int &table = 0);
+    bool setValue(const QString &group, const unsigned short int &time, const bool &noError = false, const unsigned short int &table = 0);
+    const QString value(const QString &group, const unsigned short int &table = 0);
 
     static QMap<int, bool>* getNoErrorList(const QString &fileName, const QString &mode);
-    static bool hasNoErrorTrue(const QString &fileName, const QString &mode, const int &table);
+    static bool hasNoErrorTrue(const QString &fileName, const QString &mode, const unsigned short int &table);
     static bool isAllTableWithNoErrorTrue(const QString &fileName, const QString &mode);
     static bool isAllTableWithNoErrorFalse(const QString &fileName, const QString &mode);
